Fix CollisionLayer DistanceBetween bugs and make more jobs BurstSchedulable
DistanceBetween using CollisionLayer didn't inflate the initial AABB collider. And none of the variants accounted for negative maxDistance values. This is all fixed now.

In addition, I did a project-wide search to find more instances of IJob, IJobFor, and IJobParallelFor and switched them to their BurstSchedulable counterparts. I'm likely going to have to undo that when 1.0 drops, but whatevs. That's the least of my 1.0 concerns.